kurs/lab_31/my_project/requirements.txt
Django==3.2.6
gunicorn==20.1.0
psycopg2-binary==2.9.1
django_filter==22.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2

